Add migration to Cloud Profile adding docker runtime to all MachineImageVersions (#4500)
* Inject logger into CloudProfile Controller

Such that it can properly be tested

* Make error log output more verbose

Such that you understand which reference is preventing the deletion of a Cloud Profile from happening

* Add CloudProfile controller tests

Shamelessly copied over from ControllerDeployment

* Fix import formatting

* Add migration to Cloud Profile adding docker runtime

Previously, docker was implicitly supported by each MachineImageVersion. In order to simplify the rest of the codebase and to allow producing images without docker support in the future, we require an explicit definition of `docker` for each MachineImageVersion. In order to relieve operators from the manual and error-prone task of doing this by hand, we're migrating Cloud Profiles for them automatically.

Note: Until this commit is reverted, it is impossible to add a MachineImageVersion without docker support!

* remove unnecessary block checking for nil