Fix race in WaitUntilExtensionObjectsDeleted
Remove filtering for resources with a deletion timestamp.
This check races with the cache: the (unwanted) resources are deleted only a few
moments earlier and thus, the cache might not have observed the addition of
the deletion timestamp, when WaitUntilExtensionObjectsDeleted checks it.
If that happens, the func will not wait for the deleted resources to be cleaned up.
Instead of checking the deletion timestamp in WaitUntilExtensionObjectsDeleted,
components should pass a proper predicate filtering objects to wait for by name.