Implement WaitCleanupStaleResources for container runtimes
Before, WaitCleanup was used for two different purposes: waiting for cleanup
of all resources after deleting them via Destroy and waiting for cleanup of
stale (unwanted) resources after deleting them via DeleteStaleResources.
Filtering for wanted or unwanted objects was done inside
WaitUntilExtensionObjectsDeleted by checking the deletion timestamp of objects.
But this races with the cache, so the filtering needs to be done on the component's
side by name.
For this, WaitCleanupStaleResources takes the same approach like DeleteStaleResources.