Remove docker-specific kubelet flags when container runtime is set to containerd (#4693)
* Remove docker-specific kubelet flags for containerd

These flags were never meant to be set with containerd and don't do anything when you set them with anything else than 'docker': https://kubernetes.io/docs/reference/command-line-tools-reference/kubelet/

* Remove duplicative computation of cliflags

This was essentially a re-implementation of production code in the test. Ideally, the component test wouldn't care about the actual CLIFlags as this is already tested in cliflags_test.go

* Re-order cliflags for kubelet

Such that it easier to see which flags are added under which conditions

* Re-order imports

* Validate imagePullProgressDeadline is docker-only

Users can set imagePullProgressDeadline globally in Shoot.spec.kubernetes.kubelet or for an individual worker. We now throw a validation error when this setting has no effect. This is true when
* The flag is set globally, but there is no worker pool configured using docker
* The flag is set on an individual worker, but this worker is not using docker