Remove creation of static node bootstrapToken
This commit is part of issue [#3898](https://github.com/gardener/gardener/issues/3898)
which replaces the long-valid bootstrap-token shared between nodes with a
short-lived token unique for each node.

The new flow of using smaller-scoped, short-lived tokens was already active
once you updated to compatible versions of the infrastructure-extension,
operatingsystem-extension and of gardener/gardener as specified in:
https://github.com/gardener/gardener/issues/3898

With this commit we are now removing the old secret from the Shoot
which means you need to run supported versions of the os-extensions
and the infrastructure-provider-extensions when
upgrading gardener to this version.