Replace cluster specific envoy filter for reversed vpn with a static one. (#4434)
* Replace cluster specific envoy filter for reversed vpn with a static one.

As described in #4381, the previous approach with cluster specific envoy filters causes connections to be reset
when a new cluster is created or an existing one is deleted, i.e. in case a reconfiguration happens on the istio
ingress gateway/envoy proxy. In this change the per cluster approach is replaced with a single static envoy filter
so that no reconfigurations happen on cluster creation/deletion.

* add external authorization for envoy

* Clean up old resources properly: Gateway, VirtualService and EnvoyFilter are now removed on update.

* add tests for external authz server

* Fixed tests after removal of old gateway, virtual service and envoyfilter.

* Improved doc string, removed unnecessary constant and used constant for default domain.

* Add support for ExposureClasses in conjunction with new reversed vpn architecture.

* remove interface from ext authz server

* rename import alias to gardencorev1beta1

* add pdb and antiaffinity to reversed-vpn-auth-server

Co-authored-by: Sebastian Stauch <sebastian.stauch@sap.com>