Drop finalizer and `ServiceAccount` deletion handling from token requestor
Initially, we have added the finalizer since the controller is creating resources (`ServiceAccount`s) in the target cluster. However, in real-world scenarios, this brings only little to none benefit since we will set the 'skip-deletion' annotation for most of the shoot token secrets anyways, i.e., there is nothing to clean up when the `Secret` in the source cluster gets deleted.
Even worse, if we kept the finalizer then we would need to introduce additional logic for explicitly deleting some of the control plane components during shoot migration or shoot deletion (today, we rely on the recursive cleanup performed by the namespace-controller when the shoot `Namespace` is deleted).

Hence, for pragmatic reasons, let's remove the finalizer and the `ServiceAccount` deletion handling now.

Introducing a more sophisticated solution (watching `ServiceAccount`s in the target cluster) is non-trivial and would result in a lot of additional code/complexity.