Remove bootstraptoken file from original osc (#4994)
* Remove bootstraptoken file from original osc

Without this change the original osc contains all files and units
from the downloader osc. This means that a valid bootstraptoken
replaced by the MCM in the userData will be overwritten with a
token placeholder when the cloud-config is downloaded.

This becomes relevant in debugging scenarios where the node did not
start up correctly (for whatever reason). Then the bootstrap token is
overwritten and the node can never join. Also it adds confusion when
debugging the node as it seems the token placeholder was never replaced.

Co-authored-by: Daniel Foehr <daniel.foehr@sap.com>

* append downloader files in range loop

Co-authored-by: Daniel Foehr <daniel.foehr@sap.com>