Disable parallel execution of update-codegen (#5020)
* Disable parallel execution of update-codegen

We sporadically saw errors when running check-generate
locally or in pipelines. Appearance of those errors correlate
with the time we introduced parallelization for `make generate`.

When looking at the logs of a failing run we can see that the first
error appears when running update-codegen.sh. Consequently, this
removes the parallel execution in this script.

We were not able to consistently reproduce this issue to find the exact
root cause.

Related issue: https://github.com/gardener/gardener/issues/4570

* Include landscapercontrolplane_groups in parallel generation

Co-authored-by: Tim Ebert <tim.ebert@sap.com>