Capture a metric for a persistent volume issue
We observed that sometimes persistent volumes are not mounted properly
and the local disk is used instead. To detect that situation, we compare
2 metrics:
- the size of the persistent volume claim, as exposed by kube-state-metrics
- the size of the mounted volume, as exposed by the kubelet

Due to the overhead of the filesystem, they are not exactly the same
(filesystem capacity is about 2% less than the raw disk capacity).

We expose these metrics for federation in the garden cluster, where an
alert could be defined to notify the operators.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>