Fixed a couple of panels in the istio mesh dashboard and exposed the istio http metrics. (#5077)
Some of the used performance counters were removed from istio quite some time ago.
Apparently, nobody noticed the failing dashboards. Now, at least the dashboards
for virtual services, destination rules and gateways should work again.
Furthermore, a dashboard with a resource that is no longer available was replaced
with workload entries, which we currently do not use.

There were also a couple of panels with incorrect reporter values and one panel
with incorrect tag names.

Last but not least, the additional envoy filters expose the http metrics of istio
so that corresponding dashboards also receive values. The data is currently only
relevant for the reversed-vpn as the apiserver connections go through tcp.