provider-local: Add `coredns` deployment for in-cluster DNS
Note: This allows the components running on shoot worker nodes (in the machine pods) to resolve the `*.local.gardener.cloud` DNS names such that they can communicate to the respective shoot's `kube-apiserver`. In an earlier commit, we had added the `machinepod` webhook which was configuring exactly this `coredns` server for the machine pods.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>