Local development: Make sure gardenlet generates client cert
Initially introduced with https://github.com/gardener/gardener/pull/5086, this PR simplifies the client certificate generation by simply creating a `CertificateSigningRequest` instead of creating the certificate ourselves. In addition, this procedure is now enabled also for the kind-based development setup.

This also allows to drop the existing `gardenlet.key` as well as the `.gitignore` files since the kubeconfigs will be generated ad-hoc and not persisted.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>