🛡 Eliminate client certificate for `cloud-config-downloader` (#5121)
* Minor cosmetics

* Remove common parts of `Deploy` and `Restore` into new `reconcile` method

* Create shoot access secret for `cloud-config-downloader`

* Adapt `cloud-config-downloader` script for downloading its own shoot access token

Earlier (before this commit), the `cloud-config-downloader` had only read the `cloud-config` secret which contains the user data for configuring the respective node. It had used a client certificate for this which was never rotated/changed.
Now, it first downloads the `cloud-config-downloader` secret in the `kube-system` namespace of the shoot cluster. This contains its own token for reading the `cloud-config` secret.

There are three cases:

1) A new node is created -> In this case, the bootstrap token for the kubelet is used to read the `cloud-config-downloader` secret.
2) An existing node is updated but it does not have the new shoot access token yet -> In this case, the old client certificate is used to read the `cloud-config-downloader` secret. Afterwards, these credentials are permanently deleted from the disk.
3) An existing node is updated and it has the new shoot access token yet -> In this case, the shoot access token is used to read the `cloud-config-downloader` secret.

In all scenarios, the token is stored on disk and updated whenever the content of the `cloud-config-downloader` secret changes (after `24h`).

* Address PR review feedback