Fix `Or` predicate
Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

The upstream `predicate.Or` function does not support injecting dependencies into the "sub"-predicates. This caused a `nil` pointer exception in the `dnsrecord` controller since the needed dependencies were not injected into the `ShootNotFailed` predicate.
Tim E. will open an upstream PR at `kubernetes-sigs/controller-runtime` to fix this for both the `And` and the `Or` functions. Until this is merged, released and revendored we have to live with our own copy of this function.