Refactor extension controller predicates (#5297)
* Introduce default controller predicates

This PR introduces a function `DefaultControllerPredicates` which can be called by the individual controller packages (see later commit) as part of their `DefaultPredicates` functions. Today, this is duplicated n-times (once in each controller package).
Moreover, instead of glueing different predicates together in a unreadable, unmaintainable, and badly testable way, the introduced `defaultControllerPredicate` encapsulates the logic of the earlier "glued" predicates.

For example, the `DefaultPredicates` function in the `infrastructure` controller packages looks as follows:

```
func DefaultPredicates(ignoreOperationAnnotation bool) []predicate.Predicate {
	if ignoreOperationAnnotation {
		return []predicate.Predicate{
			predicate.GenerationChangedPredicate{},
			extensionspredicate.ShootNotFailed(),
		}
	}

	return []predicate.Predicate{
		predicate.Or(
			predicateutils.HasOperationAnnotation(),
			extensionspredicate.LastOperationNotSuccessful(),
			predicate.And(
				predicate.GenerationChangedPredicate{},
				extensionspredicate.IsDeleting(),
			),
		),
		extensionspredicate.ShootNotFailed(),
	}
}
```

There have been many attempts in the past months and years to maintain this, fix bugs, workaround issues, etc. However, it is not well maintainable, untested, and has potential for further regressions in the future.
The new `defaultControllerPredicate` aims at overcoming these downsides.

The `DefaultControllerPredicates` introduces the concept of "precondition predicates". We consider the `ShootNotFailed` and the `IsInGardenNamespace` predicates as such preconditions that might or might not be relevant for the individual extension controllers.
For example, the `BackupBucket`, `BackupEntry`, and `Bastion` controllers are operating on global resources not directly associated to any `Shoot`. Hence, they wouldn't active both of these preconditions.
The `DNSRecord` controller is the only one operating on a resource which may also reside in non-shoot namespaces. Hence, it would active both of them.
All other controllers would only active the `ShootNotFailed` precondition since they are related to a `Shoot` and only deployed in shoot namespaces.
The next commit adapts the individual controller packages.

On the way, when reviewing the changes you might notice that the `defaultControllerPredicate` and the preconditions are not a 1:1 adaptation of the glued predicates from before.
For example, the "is deleting" case looks different now since it working as expected in all cases before. More concretely, extension resources were not picked up again in case they were deleted and the deletion operation failed before the controller restarted.
This is now overcome with the "statusChanged" condition in the `UpdateFunc`. It ensures that controllers pick up such resources while preventing that they can trigger themselves endlessly when updating the `status`es.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Adapt `DefaultPredicates` functions of extension controllers

* Drop old/no longer needed code

* Drop unused predicates

* Move predicates related to GRM to `resourcemanager` folder

* Address PR review feedback

* Fix `Or` predicate

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

The upstream `predicate.Or` function does not support injecting dependencies into the "sub"-predicates. This caused a `nil` pointer exception in the `dnsrecord` controller since the needed dependencies were not injected into the `ShootNotFailed` predicate.
Tim E. will open an upstream PR at `kubernetes-sigs/controller-runtime` to fix this for both the `And` and the `Or` functions. Until this is merged, released and revendored we have to live with our own copy of this function.

Co-authored-by: Tim Ebert <tim.ebert@sap.com>