Fix failing creation of hibernated shoots (#5300)
* Make `Replicas` an `int32` pointer

This is to be able to identify whether the `Replicas` field in the `Values` was already set or whether it is uninitialized. For example, in shoot deletion the flow already sets the replicas field.

* Make `botanist.determineKubeControllerManagerReplicas` function reusable

* Compute correct replicas when deploying GRM

The replicas are now computed similar to when they are computed for the KCM deployment. Since the hibernation/creation/etc. cases are already properly handled in the `determineControllerReplicas` function, it was possible to no longer check for hibernation in the `mustBoostrapGardenerResourceManager` function

* Do not deploy MCM with 1 replica when shoot is hibernated

Earlier, MCM was deployed with 1 replica and scaled down again afterwards. This is (a) unnecessary and (b) does not work anymore when MCM is using a token managed by token requestor.
By calling `kutil.CurrentReplicaCountForDeployment` this is avoided.