Switch update strategy for Shoot conditions
Without this it can happen that the shoot conditions controller acts
upon a stale shoot object leading to seed conditions only being partially
updated in the shoot resource (timestamps are updated, status and message are not updated).

Prerequisite:
* Cache with Shoot object is stale (for example due to overloaded/unavailable gardener-apiserver)
* Condition in Seed flips during the overload/apiserver unavailability for example
  from status "True" to "False" and back to "True".

As explained in gardener/gardener#5295 this will lead to updates only for `LastTransitionTime` and `LastUpdateTime`
in the Shoot as the StrateticMergePatch contains no changes to status and message.
The Shoot object would then never be enqueued again unless the Seed conditions change again or the GCM is restarted.

This commit fixes this problem as now the shoot resource must be up-to-date (optimistic locking)
otherwise the reconcilation will be retried.