Better handle stale caches in conditions controller
In case of the Seed reflector loosing connection and needing to do a full relist the
informer emits `Add` events for all objects in the cache. In this case we want
the controller  to enqueue them again, because it might have missed
relevant `Update` events.

As a safety net we also want to enqueue in case of periodic resync events.