Simplify and cleanup things in test framework
Some of the changes are required for the e2e test, e.g. to allow instanciating
the shoot from code instead of from CLI flags.
Some other changes cleanup and simplify stuff to make it as easy and short as
possible to call the different steps implemented in the framework, e.g. creating
a shoot, in order to minimize code duplication in testmachinery and e2e tests.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>