Upgrade to ginkgo v2 (#5313)
* Upgrade to ginkgo v2.1.0

* Adapt ginkgo import paths

* Upgrade to gomega v1.18.0

* Drop obsolete table extension imports

* [automated] make revendor

* Replace deprecated done channel

* Remove AfterSuite-specific timeout

* Fix broken table tests

* Adapt custom reporter to changed reporting infrastructure

This is only a temporary adaption, as we are using the temporary
"compatibility layer", which will be removed again soon.
Hence, we have to rework this and adapt for the long-term.

* [automated] make revendor

* Replace wrongful {Before,After}Suite usages

* Fix tests that didn't verify mock calls so far

These suites were missing a call to gomock.Controller.Finish(), which verifies
calls made to mocks. With ginkgo v2, support in GinkgoT() was extended and now
Finish doesn't have to be called explicitly anymore. Now, these suites started
to fail, as there were missing mock calls or calls with unexpect parameters.

* Replace CurrentGinkgoTestDescription() with CurrentSpecReport()