Add import-boss to restrict imports (#5347)
* Rework testing docs, split from dependency management

* Document motivation of import restrictions

Also, run `go mod tidy` before `go mod vendor` to make `make revendor`
deterministic.

* Add import-boss and run it on make check

* Restrict imports in pkg

* Restrict imports in extensions

* Move DNSRecord name constants to extensions API

This is done to eliminate the dependency from
extensions/controller/dnsrecord to pkg/operation.
Also, differentiate between names for DNSRecords and legacy dns
objects.

* Move utils using listers to plugin

This is done to eliminate the dependency from pkg/utils/gardener
to the generate clientsets under pkg/client.
Generated informers/listers are only used in admission plugins,
hence plugin/utils is a good fit for these helper functions.

* Move gardenlet-specific filter funcs to controllers

This is done to eliminate the dependency from pkg/controllerutils
to pkg/gardenlet.

* Drop unused scheduler helpers in test framework

This is done to eliminate the dependency from test/framework
to pkg/scheduler.

* Move validation helpers to dedicated packages

This is done to allow importing the logic in extensions without pulling
in all dependendies of the validation package. Actually, other projects
should never depend API validation packages. But we don't have a sync
out mechanism which filters out validation code, so here we are.

* Move image name constants to dedicated package

This is done to eliminate the dependency from extensions to charts.

* Restrict imports in cmd/gardener-apiserver

* Restrict imports in pkg/apis

* Move timewindow utils to dedicated package

This is done to minimize the blast radius of importing the API packages,
as the pkg/utils package pulls in a lot of other stuff.
A few other imports of utils functions are replaced as well.

* Tighten import restrictions in pkg/api

* Move dependency to external-dns-management to validation package

This is done to not force external dependents to depend on
external-dns-management as well.

* Move DNSRecord name constants to v1beta1constants

as per suggestion by rfranzke

* Tighten import restrictions in pkg/apis/seedmanagement

* Move log level/format consts into component API pkgs

This is done to eliminate the dependency from the gardenlet config
API package to pkg/logger. This dependency was pullin in zap, logrus
and controller-runtime.

* Use literals instead of constants in component API pkgs

This is done to eliminate the dependency from the gardenlet config
API to client-go. This caused the seedmanagement API to pull in
unwanted dependencies that required pinning client-go and k/api.

* Adapt Dockerfile.local

This makes use of the trimmed down package dependencies and adapts
to the restructuring.
Also, build up to 3 images in parallel in skaffold.yaml

* Address review suggestions

* Restrict imports in component config APIs

Although they are probably not reused frequently in other projects,
it's still a good idea to keep them isolated similar to other API
packages. This will also help us to cut down transitive dependencies
of the seedmanagement API package.

* Use dedicated scheme in controllermanager config

This is done to eliminate the dependency to client-go, that is
not needed in this case.