Drop Shoot ConfigMap controller (#5392)
* Drop Shoot ConfigMap controller

* Stop setting resourceVersion field in Shoot

Keeping this in the ShootSpec will be confusing, as it basically has
no effect anymore.
Also, stop validating the ConfigMap on updates to the Shoot, because
intermediate updates to the ConfigMap will be validated already by the
webhook. Only validate the referenced ConfigMap on Shoot creation and
updates, that add the reference. In that case, the ConfigMap might not
have been validated yet.

* Move AuditPolicy webhook to ValidatingWebhookConfiguration

as it is no longer mutating.

* Unset resourceVersion field in shoot spec

Use the next shoot maintenance to unset the field to prevent confusing
users by a field, that has no effect anymore.