Consider all `ManagedResource`s for system component care checks
Traditionally, we have added the name of the managed resources to check to the `managedResources` list. Then we read each `ManagedResource` (`n` calls to the kube-apiserver instead of `1` list call, but anyways) and (implictly) checked whether it existed and whether it is healthy. See for example what was done when the `coredns` component was introduced: https://github.com/gardener/gardener/pull/4536/commits/f1e91f4ba411409f3b2bc941fba81cdcb34acb2c

Now with the `kube-proxy` component, the names of the `ManagedResource`s are no longer static and computing them is "expensive". Hence, we cannot continue like this anymore and so I adapted the code such that all `ManagedResource`s by gardenlet are considered and checked (all those with `origin=gardener` label).