Fix the request latency panel
The panel used an incorrect query before: it applied the
rate function to the apiserver_latency_seconds:quantile metric.

The apiserver_latency_seconds:quantile is the average latency of the
requests, in seconds - the rate function is not meaningful in that case.

The latency values can not be summed up in a meaningful way,
but we can pick the maximum value for plotting on the panel.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>