Fix overflowing float to int32 conversion
ValidateNodeCIDRMaskWithMaxPod calculates the ip addresses available
by doing a float 2**(32-nodeCIDRMaskSize) and casting that to an int32.
Unfortuantely, for nodeCIDRMaskSize 0 (or 1) which is provided in a
unit test this value isn't representable as a int32 and it hits
architecture specific behavior that results in it being a large negative
number on x86_64 but a large positive number of arm64.  The negative
number is intended the following if condition passes while it wouth
otherwise have failed.

This change uses a uint32 in the two places where Pow() is used
to not be subject a the 32b overflow and enables make test to complete
successfully on an arm64 system.

Signed-off-by: Ali Saidi <alisaidi@amazon.com>