[GEP-18] Introduce new secrets manager and adapt CAs and `kube-scheduler-server` secrets (#5503)
* Documentation for new secrets manager

* Vendor `github.com/mitchellh/hashstructure/v2`

This module will be used for computing a hash for the configuration structures of secrets.

* Add manager structure

* Add new secret config type for certificate bundle secrets

The `infodata`-related functions are not implemented on purpose since they will not be used and dropped soon, hopefully.

* Implement `GetOrGenerate`

* Implement `GetByName`

* Implement `Cleanup`

* Add fake implementation for secrets manager

* Integrate `SecretsManager` into `Botanist`

* Move wanted CA secret configs from `wanted_secrets.go` to `secrets.go`

* CAs are now generated by the new secrets manager

* Adapt legacy secrets manager to no longer generate CAs

Instead, the caller (botanist) must now provide the map of existing CAs to it. This is needed because the legacy secrets manager still generates most of the server/client certs. Until we have adapted those we need this, but it can be dropped in the future.

* Extract function from `ShootStateSync` controller for reuse

* Add new `ShootSecret` controller for persisted secrets in `ShootState`

* Add integration test for `ShootSecret`  controller

* Extend gardenlet config API for controller config and add it to factory

* Add new `labels` field to `GardenerResourceData` type

* Create secrets from `ShootState` when shoot is restored

* Enhance certificate config to optionally not publish CA cert

* `kube-scheduler` creates its server cert using `SecretsManager`

Additionally, it does no longer rely on a published `ca.crt` key in the server secret's `data`. Instead, it fetches the CA bundle secret and uses this one. This is to not duplicate the CA certificate unnecessarily.

* Address PR review feedback

* Address PR review feedback (2)

- Rename GetOrGenerate to Generate
- Rename GetByName to Get