Increased kube-proxy memory limit to eight times the size. (#5552)
Even though the vertical pod autoscaler is configured for kube-proxy there
might be intermediate states when the vertical pod autoscaler is not applying
its recommendations. For larger clusters the currently defined memory limit
might be too low. This change increases it to be on the safe side as kube-proxy
is crucial for setting up the service network and especially if externalTrafficPolicy
is set to local as it then also acts as the health endpoint for load balancers.