Drop `botanist.RewriteShootSecretsIfEncryptionConfigurationChanged` function
This function was introduced due to the following scenarios:

1) Old shoots with unencrypted etcd data will now get the encryption configuration (hence, secrets must be rewritten to become encrypted in etcd)
2) Encryption key changes/rotates
3) Secrets were forced to plain text (via `shoot.gardener.cloud/etcd-encryption-force-plaintext-secrets=true` annotation on the `etcd-encryption-secret` in the shoot namespace in the seed)

Now, 1) is no longer valid since we only support shoots with encrypted etcd data since very long. 2) is not implemented. 3) will be dropped (this feature was never used and is not too helpful anyways)

Given that the code is not optimal (sequential instead of parallel) and that it wouldn't work when https://github.com/gardener/gardener/issues/4606 gets implemented, let's drop it now. It has no unit tests anyways, so let's rather invest the time to write it properly when we either implement encryption key rotation or support end-user configuration for additional resources that should be encrypted (#4606).