Apply suggestions from review comments
Removed the additional check to forbid changing the
shoot.spec.seedName if the last operation is not successful as that
could block bad case scenario and was only added to cover a very specific
corner case.

Included a check that will forbid changes to the shoot.spec if the
shoot.spec.seedName has been changed in a previous patch/update,
but migration operation has not been started by the gardenlet yet.