[GEP-18] Adapt `etcd-encryption-secret` to new secrets manager (#5616)
* New `secrets.ConfigInterface` implementation for etcd encryption keys

* Move codec creation from `configmaps.go` to `kube_apiserver.go`

Since this is no longer restricted to content in `configmaps.go` only

* `kubeapiserver` component generates ETCD encryption key + configuration

* Keep existing credentials for backwards compatibility

* `Deployment` uses self-generated ETCD encryption config secret

Instead of the one injected via the `Secrets` from outside

* Delete legacy etcd encryption configuration state and secret

Only after successful KAPI deployment

* Drop `botanist.RewriteShootSecretsIfEncryptionConfigurationChanged` function

This function was introduced due to the following scenarios:

1) Old shoots with unencrypted etcd data will now get the encryption configuration (hence, secrets must be rewritten to become encrypted in etcd)
2) Encryption key changes/rotates
3) Secrets were forced to plain text (via `shoot.gardener.cloud/etcd-encryption-force-plaintext-secrets=true` annotation on the `etcd-encryption-secret` in the shoot namespace in the seed)

Now, 1) is no longer valid since we only support shoots with encrypted etcd data since very long. 2) is not implemented. 3) will be dropped (this feature was never used and is not too helpful anyways)

Given that the code is not optimal (sequential instead of parallel) and that it wouldn't work when https://github.com/gardener/gardener/issues/4606 gets implemented, let's drop it now. It has no unit tests anyways, so let's rather invest the time to write it properly when we either implement encryption key rotation or support end-user configuration for additional resources that should be encrypted (#4606).

* Delete no longer required code

* Address PR review feedback

* Address PR review feedback