Properly define application and runtime resources
When the cluster type is 'seed' then all resources are managed as part of a `ManagedResource`.
When the cluster type is 'shoot' then some resources ("runtime" resources) are deployed and deleted directly via the client while other resources ("application" resources) are managed as part of a `ManagedResource`.

Now that we have this proper distinction, some duplicate coding could be deleted/prevented.