[ReversedVPN]: Add several panels to identify issues with DNS cache host limit faster. (#5660)
The DNS cache of the vpn-seed-server's envoy proxy side car has a host limit. To help identify
scenarios faster, in which this limit is hit, the reversed vpn's dashboard is extended by several
panels. Now, both upstream and downstream request rates are visible. In addition to that,
the current number of hosts in the DNS cache is shown together with the amount of host overflows.