[GEP-18] Auto-renew secrets which are about to expire (#5679)
* Maintain lifetime labels for secrets

There are two new labels `issued-at-time` and `valid-until-time`. For certificate secrets their values are derived from the generated certiifcate itself. For other secrets, the `issued-at-time` is the current time.

* Add `Validity` option for `Generate`

This allows specifying the validity of a secret for non-certificate secrets (here, it is not respected since it doesn't make much sense - the validity is always deduced from the generated certificate).

* Auto-renew secrets which have reached 80% of their validity

* Drop coding for custom ingress certificate renewals

* Address PR review feedback