Move VPA CRDs from Helm chart to a Golang component (#5663)
* New `component.Deployer` for VPA CRDs

Similar to what we do for extension CRDs: https://github.com/gardener/gardener/tree/master/pkg/operation/botanist/component/extensions/crds

* Use VPA CRD deployer in seed controller

The `charts/seed-bootstrap-crds/charts/vpa` was symlinked in `charts/shoot-core/components/charts/vertical-pod-autoscaler/charts/crds`, so for now we move the YAMLs there. They will be dropped in a future commit when also the shoot deployment gets refactored.

* Make CRD deployer workable with a `ManagedResource` registry

When VPA is enabled for shoots then the CRDs are deployed as part of a `ManagedResource` (and not directly with an applier - this is only done when they are deployed to the seed).