Introduce new client CA certificate for shoots
For backwards-compatibility reasons, the client CA is the same like the cluster CA for existing clusters.
For new clusters, a CA certificate different from the cluster CA shall be generated right away. This is why we have to generate the client CA first and cannot simply add the certificate secret config to the `b.wantedCertificateAuthorities` map because the generation order is randomized. Otherwise, if the `ca` secret is generated before `ca-client` then the cluster CA certificate would be adopted (which is not what we want for new clusters).