Add SecretsManager utilities for extensions
Idea is to implement the contract for CA rotation that extensions need to fulfill
centrally once, so that extensions can work with a simple list of secret configs
without worrying about CA rotation.
This commit adds some helpers for working with a secrets manager and a simple list
of secret configs that extension controllers will need for fulfilling the contract.