Adapt generic ControlPlane actuator to secrets manager
Use a secrets manager to manage all secrets belonging to a ControlPlane object.
Extensions pass a list of wanted secret configs and the generic actuator takes
care to create, rotate, delete them according to the contract for CA rotation.
Pass on the used secrets manager to the ValuesProvider as a SecretsGetter, so
that the ValuesProvider can determine the names of generated secrets, retrieve
CA bundles and set the respective values accordingly.