[GEP-18] Introduce dedicated client CA for shoot clusters (#5779)
* Introduce new client CA certificate for shoots

For backwards-compatibility reasons, the client CA is the same like the cluster CA for existing clusters.
For new clusters, a CA certificate different from the cluster CA shall be generated right away. This is why we have to generate the client CA first and cannot simply add the certificate secret config to the `b.wantedCertificateAuthorities` map because the generation order is randomized. Otherwise, if the `ca` secret is generated before `ca-client` then the cluster CA certificate would be adopted (which is not what we want for new clusters).

* Adapt `kubeapiserver` component to use client CA

* Adapt `kubecontrollermanager` component to use client CA

KCM always uses the current client CA secret which is used to sign new client certificates.

* Adapt `kubescheduler` component to use client CA

* Sign client certificates by new client CA

* Address PR review feedback