[GEP-18] Adapt extensions library to new secrets manager (#5803)
* Document extension requirements for CA rotation

Also, explain how to reuse SecretsManager for fulfilling these requirements.
Update documentation for kubeconfigs for shoot clusters as well.

* Cleanup `useTokenRequestor` and `useProjectedTokenMount` flags in worker actuator

* Cleanup `legacySecretNamesToCleanup` in generic controlplane actuator

* Simplify NoopValuesProvider

- drop unneeded client injection
- drop unneeded pointers

* Add SecretsManager utilities for extensions

Idea is to implement the contract for CA rotation that extensions need to fulfill
centrally once, so that extensions can work with a simple list of secret configs
without worrying about CA rotation.
This commit adds some helpers for working with a secrets manager and a simple list
of secret configs that extension controllers will need for fulfilling the contract.

* Extract Get into dedicated interface

This is to pass a read-only SecretsManager to callers to prevent them
from wrongly calling other methods than Get.

* Adapt generic ControlPlane actuator to secrets manager

Use a secrets manager to manage all secrets belonging to a ControlPlane object.
Extensions pass a list of wanted secret configs and the generic actuator takes
care to create, rotate, delete them according to the contract for CA rotation.
Pass on the used secrets manager to the ValuesProvider as a SecretsGetter, so
that the ValuesProvider can determine the names of generated secrets, retrieve
CA bundles and set the respective values accordingly.

* Drop unused legacy code

* Generate some secrets in local ControlPlane controller

Secrets are actually not needed, but can be used to test the CA rotation
mechanics that were implemented in the generic ControlPlane actuator.

* Deflake CA rotation e2e test

* Verify provider-local secrets in CA rotation e2e test

* Address review suggestions

* Add client cert to local ControlPlane controller

* Nits

* Default `useProjectedTokenMount` to true in terraformer library