Use unique common names for CA certificates
Some components (like the `vpn-seed-server`/OpenVPN) require the common
names to be unique, otherwise it will fail with:

```
2022-04-25 12:28:06 Cannot load CA certificate file /tmp/ca.bundle (entry 2 did not validate)
2022-04-25 12:28:06 Cannot load CA certificate file /tmp/ca.bundle (only 1 of 2 entries were valid X509 names)
2022-04-25 12:28:06 Exiting due to fatal error
```

See:

- https://github.com/OpenVPN/openvpn/blob/23ae78e657052748be68b623ca8122e4103dc7e0/src/openvpn/ssl_openssl.c#L1614-L1623
- https://github.com/OpenVPN/openvpn/blob/23ae78e657052748be68b623ca8122e4103dc7e0/src/openvpn/ssl_openssl.c#L1654-L1664

Consequently, we decided to overwrite the common name with the (unique)
secret name for CA secrets. This has the positive side effect that it
makes examining certificate secrets easier since it's now easier to see
which CA has signed which server/client certificate.