Make #4289 backwards-compatible
In fact, we now generate two cloud-config secrets in the `kube-system` namespace for each worker pool: One with the new name (based on adapted logic of the `operatingsystemconfig.Key` function), and one with the old name.
This ensures that existing nodes can still find their cloud-config secrets and update the `download-cloud-config.sh` script so that it contains the new secret name for future invocations.

Also, we want to get rid of the introduced migration code very soon again, so we must ensure that all shoots get migrated. The regular OSC and CC-execution deployment steps are depending on the `Infrastructure` and `ControlPlane` reconciliation since some extensions might have to inject the `cloud-provider-config` for the kubelet into the OSC. However, when there are failed shoots with invalid credentials then the regular OSC deployment steps would not get executed (which would make it unsafe for us to remove it again with a future release). Hence, for migration purposes we introduce another OSC deployment step at the very beginning of the flow (only when the `ControlPlane` resource exists and is ready, otherwise extensions might not work).