Fix caBundle of istio's validating web hook.
The validating web hook configuration of istio does not contain a ca bundle. This is only injected at
runtime by istio itself. Unfortunately, the logic how to identify such validating web hooks has changed
between istio 1.10 and 1.11. Starting with 1.11 the web hooks are identified using a label selector.
The label selector uses the istio.io/rev key, which should have the revision as value. We do not set
any revision as we do not work with canary deployments at the moment. Therefore, it is empty in our
case.
With the added label, the validating web hook is properly updated with the ca bundle by istio.
Please note that it is not possible to add the revision label to the set of standard istio labels as
those labels are also used as selector in the istio pilot deployment. As selectors in deployments are
immutable, we would need to delete/recreate the deployment, which seems to be unnecessary at this time.