Deflake: should set ManagedResource to unhealthy
There was a missing To(Succeed()) when removing the test finalizer. Because
of a stale in-memory object it might have happened, that the test tried to
patch including the controller's finalizer which had already been removed
from the actual ManagedResource. Adding a finalizer to an object which is
already in deletion is denied by API server. The missing To(Succeed())
caused that such a failure would go unnoticed and the ManagedResource would
not be deleted in the end, because the test finalizer was still present.

Now, the test finalizer is removed with optimistic locking and retries to
make sure we can successfully remove it and the ManagedResource is gone in
the end.